Fixed compiler warnings.
git-svn-id: svn://scribus.net/trunk/Scribus@13144 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/pageitem.cpp
@@ -2775,7 +2775,7 @@ void PageItem::checkTextFlowInteractions(bool allItems)
 		QRectF baseRect(getBoundingRect());
 		for(int idx = (allItems ? m_Doc->Items->count()-1 : ItemNr-1); idx >= 0 ; --idx)
 		{
-			if(idx != ItemNr) // avoids itself
+			if(idx != static_cast<int>(ItemNr)) // avoids itself
 			{
 				if(m_Doc->Items->at(idx)->asTextFrame()) // do not bother with no text frames
 				{
